Move cache types to dedicated mcp_server_cache.go file
